removed maps.js and home in components
