--- src/components/submenu/Map.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-export function Map() {
-  return (
-    <div>Map</div>
-  )
-}
-
-export default Map
